Consolidate cell seg data files from parallel projects and create summary reports

Consolidate several cell seg data files, each with its own Phenotype column, into a single file with separate columns for each phenotype.

consolidate_and_summarize_cell_seg_data(
  csd_files,
  output_dir,
  update_progress = NULL,
  col_select = NULL
)

Arguments

csd_files
A list or vector of paths to cell seg data files.

output_dir
Path to a directory where the results will be saved.

update_progress
Callback function which is called with progress.

col_select
Column selection for phenoptr::read_cell_seg_data()

Value

A single data frame containing consolidated data and columns for each single phenotype, invisibly.

Details

Create a summary report for each source file and the consolidated data.

Write the consolidated data to Consolidated_data.txt in the output directory.

The individual files must all have exactly the same Sample Name or Annotation ID and Cell ID columns. split_phenotypes is called to split the Phenotype columns.