Merge cell seg data files from parallel projects and create summary reports

Calls merge_and_split_cell_seg_data to merge the supplied cell seg data files. Additionally calls write_summary_report to create a summary report for each source file and the merged data.

merge_and_summarize_cell_seg_data(csd_files, output_dir,
  update_progress = NULL)

Arguments

csd_files
A list or vector of paths to cell seg data files.

output_dir
Path to a directory where the results will be saved.

update_progress
Callback function which is called with progress.

Value

A single data frame containing merged data and columns for each single phenotype, invisibly.

Details

Writes the merged data to Consolidated_data.txt in the output directory.

See also