Run the spatial map viewer with the given data file and export directory.

See addin_35_spatial_map_viewer() for a GUI front-end to this function.

spatial_map_viewer(csd_path, export_path)

Arguments

csd_path
Path to a Consolidated_data.txt file from the consolidation app or nearest_neighbors.txt or count_within.txt file created by the analysis app.

export_path
Path to a directory containing composite and component images for the fields in the data file.

Value

None; starts the viewer app